requirements.txt
absl-py==0.12.0
aioice==0.7.5
aiortc==1.2.0
albumentations==0.5.2
alembic==1.4.1
altair==4.1.0
antlr4-python3-runtime==4.8
appdirs==1.4.4
appnope==0.1.2
argon2-cffi==20.1.0
astor==0.8.1
async-generator==1.10
asyncio==3.4.3
attrs==20.3.0
av==8.0.3
backcall==0.2.0
base58==2.1.0
bcrypt==3.2.0
bleach==3.3.0
blinker==1.4
CacheControl==0.12.6
cachetools==4.2.1
cachy==0.3.0
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
cleo==0.8.1
click==7.1.2
clikit==0.6.2
cloudpickle==1.6.0
crashtest==0.3.1
crc32c==2.2
cryptography==3.4.7
cycler==0.10.0
Cython==0.29.22
databricks-cli==0.14.3
decorator==4.4.2
defusedxml==0.7.1
distlib==0.3.1
dnspython==2.1.0
docker==4.4.4
efficientnet-pytorch==0.6.3
entrypoints==0.3
filelock==3.0.12
Flask==1.1.2
future==0.18.2
gitdb==4.0.7
GitPython==3.1.14
google-auth==1.28.0
google-auth-oauthlib==0.4.4
greenlet==1.0.0
grpcio==1.36.1
gunicorn==20.1.0
html5lib==1.1
hydra-core==1.0.6
idna==2.10
imageio==2.9.0
imgaug==0.4.0
importlib-metadata==3.10.0
importlib-resources==5.1.2
iniconfig==1.1.1
ipykernel==5.5.3
ipython==7.22.0
ipython-genutils==0.2.0
ipywidgets==7.6.3
itsdangerous==1.1.0
jedi==0.18.0
Jinja2==2.11.3
joblib==1.0.1
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.12
jupyter-console==6.4.0
jupyter-contrib-core==0.3.3
jupyter-contrib-nbextensions==0.5.1
jupyter-core==4.7.1
jupyter-highlight-selected-word==0.2.0
jupyter-latex-envs==1.4.6
jupyter-nbextensions-configurator==0.4.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
keyring==21.8.0
kiwisolver==1.3.1
lockfile==0.12.2
lxml==4.6.3
Mako==1.1.4
Markdown==3.3.4
MarkupSafe==1.1.1
matplotlib==3.4.1
mistune==0.8.4
mlflow==1.15.0
msgpack==1.0.2
munch==2.5.0
nbclient==0.5.3
nbconvert==6.0.7
nbformat==5.1.3
nest-asyncio==1.5.1
netifaces==0.10.9
networkx==2.5.1
nose==1.3.7
notebook==6.3.0
numpy==1.20.2
oauthlib==3.1.0
omegaconf==2.0.6
opencv-python==4.4.0.46
opencv-python-headless==4.5.1.48
openvino==2021.3.0
packaging==20.9
pandas==1.1.5
pandocfilters==1.4.3
paramiko==2.7.2
parso==0.8.2
pastel==0.2.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.2.0
pkginfo==1.7.0
pluggy==0.13.1
#poetry==1.1.5
#poetry-core==1.0.2
pretrainedmodels==0.7.4
prometheus-client==0.10.0
prometheus-flask-exporter==0.18.1
prompt-toolkit==3.0.18
protobuf==3.15.7
psutil==5.8.0
ptyprocess==0.7.0
py==1.10.0
pyarrow==3.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pydeck==0.6.1
pyee==8.1.0
Pygments==2.8.1
pylev==1.3.0
pylibsrtp==0.6.8
PyNaCl==1.4.0
pyparsing==2.4.7
pyrsistent==0.17.3
pytest==6.2.3
python-dateutil==2.8.1
python-editor==1.0.4
pytz==2021.1
PyWavelets==1.1.1
PyYAML==5.4.1
pyzmq==22.0.3
qtconsole==5.0.3
QtPy==1.9.0
querystring-parser==1.2.4
requests==2.25.1
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
rsa==4.7.2
scikit-image==0.18.1
scikit-learn==0.24.1
scipy==1.6.1
scp==0.13.3
seaborn==0.11.1
segmentation-models-pytorch==0.1.3
Send2Trash==1.5.0
Shapely==1.7.1
shellingham==1.4.0
six==1.15.0
smmap==4.0.0
SQLAlchemy==1.4.5
sqlparse==0.4.1
streamlit==0.79.0
streamlit-webrtc==0.8.1
tabulate==0.8.9
tensorboard==2.4.1
tensorboard-plugin-wit==1.8.0
terminado==0.9.4
testpath==0.4.4
threadpoolctl==2.1.0
tifffile==2021.3.31
timm==0.3.2
toml==0.10.2
tomlkit==0.7.0
toolz==0.11.1
torch==1.6.0
torchsummary==1.5.1
torchvision==0.7.0
tornado==6.1
tqdm==4.60.0
traitlets==5.0.5
typing-extensions==3.7.4.3
tzlocal==2.1
urllib3==1.26.4
validators==0.18.2
virtualenv==20.4.3
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==0.58.0
Werkzeug==1.0.1
widgetsnbextension==3.5.1
zipp==3.4.1
